refactor(ipc): pragent run 队列归入 Agent 领域
pragent:*（pr-agent run 执行层：触发 / 取消 / 队列 / 历史）本是「在 PR 上跑 AI 评审工具」，
与 agent 编排同源（agent:run / AutoPilot 也经同一 run 队列派发），而非通用 PR CRUD。
故从 PR 域整体移到 Agent 域，三层一致：

- 契约 @meebox/ipc：pragent:run/listRuns/getRun/clearRuns/cancel/queue 从 PrChannels 移到
  AgentChannels（ReviewRun/PragentRunInfo import 随迁）；IpcChannels 为交集，renderer 零改动
- controller：runPragent/cancelPragent/getQueue/listRuns/getRun/clearRuns 从 pr.ts 移到 agent.ts，
  run 相关 poller import 随迁
- 注册 ipc.ts：6 个 ipcMain.handle('pragent:…') 从「PR 操作」组移到「Agent 交互」组

纯结构性重构，不改运行行为；lint / typecheck / build 全绿。
